Finished Main screen list of recipes, added image loading and saving and added a real database system

+++ app/src/main/java/ca/uottawa/leagueofsmiles/cookhelper/data/Repository.java
@@ -13,7 +13,7 @@ public interface Repository {
 
     void saveRecipe(Recipe recipe);
 
-    Recipe getRecipe(int recipeId);
+    Recipe getRecipe(long recipeId);
 
-    void deleteRecipe(int recipedId);
+    void deleteRecipe(long recipedId);
 }
